Initial working implementation of a UDP port.
